added mockable createAccount callback to decompose Persona component for easier testing
