+++ packages/mobile/__mocks__/persona-helpers.ts
@@ -1,0 +1,13 @@
+import { Dispatch } from 'redux'
+import * as React from 'react'
+
+export async function createAccount(
+  getPrivateKey: () => Promise<string>,
+  dispatch: Dispatch,
+  personaAccountCreated: boolean,
+  accountAddress: string | null,
+  setPersonaAccountCreated: React.Dispatch<React.SetStateAction<boolean>>,
+  tag: string
+) {
+  // yawn
+}
