+++ packages/mobile/src/account/Persona.tsx
@@ -1,6 +1,5 @@
 import Button, { BtnSizes, BtnTypes } from '@celo/react-components/components/Button'
 import { generateKeys } from '@celo/utils/lib/account'
-import { serializeSignature, signMessage } from '@celo/utils/lib/signatureUtils'
 import * as React from 'react'
 import { useCallback, useState } from 'react'
 import { useAsync } from 'react-async-hook'
@@ -8,14 +7,14 @@ import { useTranslation } from 'react-i18next'
 import Inquiry, { InquiryAttributes } from 'react-native-persona'
 import { useDispatch, useSelector } from 'react-redux'
 import { KycStatus } from 'src/account/reducer'
-import { showError } from 'src/alert/actions'
-import { ErrorMessages } from 'src/app/ErrorMessages'
+
 import { getStoredMnemonic } from 'src/backup/utils'
 import { readOnceFromFirebase } from 'src/firebase/firebase'
 import networkConfig from 'src/geth/networkConfig'
 import { Namespaces } from 'src/i18n'
 import Logger from 'src/utils/Logger'
 import { accountAddressSelector } from 'src/web3/selectors'
+import { createAccount } from './persona-helpers'
 
 const TAG = 'PERSONA'
 
@@ -74,36 +73,18 @@ const Persona = ({ kycStatus }: Props) => {
     return keys?.privateKey
   }
 
-  useAsync(async () => {
-    if (!personaAccountCreated) {
-      if (!accountAddress) {
-        Logger.error(TAG, "Can't render Persona because accountAddress is null")
-        return
-      }
-      const privateKey = await getPrivateKey()
-      const signature = signMessage(
-        `post /account/create ${JSON.stringify({ accountAddress })}`,
-        privateKey,
-        accountAddress
-      )
-      const serializedSignature = serializeSignature(signature)
-
-      const response = await fetch(`${networkConfig.inhouseLiquditiyUrl}/persona/account/create`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          authorization: `Valora ${accountAddress}:${serializedSignature}`,
-        },
-        body: JSON.stringify({ accountAddress }),
-      })
-
-      if (response.status === 201 || response.status === 409) {
-        setPersonaAccountCreated(true)
-      } else {
-        dispatch(showError(ErrorMessages.PERSONA_ACCOUNT_ENDPOINT_FAIL))
-      }
-    }
-  }, [])
+  useAsync(
+    () =>
+      createAccount(
+        getPrivateKey,
+        dispatch,
+        personaAccountCreated,
+        accountAddress,
+        setPersonaAccountCreated,
+        TAG
+      ),
+    []
+  )
 
   return (
     <Button
